Install Maven artifacts under "m2_repo"
Maven artifacts (plugins) are to be installed in the "m2_repo" directory
under the Embulk home directory, not directly in the Embulk home.

See EEP-4: System-wide Configuration by Embulk System Properties and Embulk Home
https://github.com/embulk/embulk/blob/master/docs/eeps/eep-0004.md

This change is to follow the convention.

"m2_repo" can be configured by the "m2RepoRelative" notation in "installEmbulkRunSet".
The default "m2_repo" is "lib/m2/repository".